Client: Create Products component and move the fetching of products etc. there, from App.js. Create Page component to wrap all page/view components in -  Page is a styled component setting margin-top to heightOfNav (exported form NavBar.js) and width so that all pages will have the same width.
