+++ client/src/utils/styling/Page.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+import { heightOfNav } from '../../components/NavBar';
+
+const Page = styled.section`
+  width: 96%;
+  margin-top: ${heightOfNav}px;
+`;
+
+export default Page;
